refactor(#67): 서비스 클래스를 서비스 인터페이스를 사용하도록 변환
